git-diff: fix incorrect comment
Previously, a comment in cmd_diff stated that if one tree-ish and no
blobs were provided, it would provide a diff between the tree and the
cache. This is incorrect because a diff happens between the tree-ish and
the working tree. Update the comment to correctly state that it is a
diff against the working tree.

Signed-off-by: Denton Liu <liu.denton@gmail.com>