git-diff: warn on git diff A..B form
git diff used to consider

    git diff <commit> <commit>

and

    git diff <commit>..<commit>

to be equal counterparts. However, rev-list-ish commands also use the
<commit>..<commit> notation, but in a logically conflicting manner which
was confusing for some users (including me!).

Deprecating the notation entirely is not really an option because it
would be an arduous process without much end-value. In addition, there
are some valid use-cases that we don't want to break.

Print a warning in the diff header when using the second form so that
users will be aware if they are confusing the forms.

Signed-off-by: Denton Liu <liu.denton@gmail.com>