branch: teach --set-remote-to option
In git-branch, the `branch.<name>.remote` variable is set alongside the
`branch.<name>.merge` variable when an upstream is configured using
`--set-upstream-to`. However, if a user wishes to change their remote
but not the merge ref, they must run
`git branch --set-upstream-to=<newremote>/<oldmerge>` which is not
convenient because they have to specify redundant information (i.e.
<oldmerge>). They could also manually edit their config but it would be
nice to provide users with an option for convenience.

Teach git-branch the `--set-remote-to` option which changes the
`branch.<name>.remote` variable so that users can change the remote of a
branch conveniently.

Signed-off-by: Denton Liu <liu.denton@gmail.com>