branch: teach --set-push-remote-to
In git-branch, the `branch.<name>.remote` variable is used to specify
the remote to fetch from and push to. The `branch.<name>.pushRemote` can
be used to override this setting so that another push remote is used.
However, users have to manually edit their configuration in order to
leverage this and the `branch.<name>.pushRemote` variable isn't very
discoverable for most Git users as it is only documented in the branch
config variables.

Teach git-branch the `--set-push-remote-to` option which changes the
`branch.<name>.pushRemote` variable so that users can override the push
remote of a branch conveniently. This also makes the configuration of a
push remote more discoverable for users since the command-line option
will appear at the top of the manpage.

Signed-off-by: Denton Liu <liu.denton@gmail.com>