branch: teach --set-remote-save-to-push
A use case for this is, if a user clones a (e.g. GitHub) fork of a
repository, the master's remote will be the user's fork. However, if a
user wants to mirror upstream's branch to origin, they'll have to modify
their configuration in two places: changing the `branch.<name>.remote`
to upstream and changing `branch.<name>.pushRemote` to the original
value of the former.

Teach git-branch the --set-remote-save-to-push option so that this can
be done in a single step.

Signed-off-by: Denton Liu <liu.denton@gmail.com>