apply: tell user location of corrupt patch file
When `git am` runs into a corrupt patch, it'll error out and give a
message such as,

	error: corrupt patch at line 87

Casual users of am may assume that this line number refers to the <mbox>
file that they provided on the command-line. This assumption, however,
is incorrect. The line count really refers to the
.git/rebase-apply/patch file genrated by am.

Teach am to print the location of corrupt patch files so that users of
the tool will know where to look when fixing their corrupt patch. Thus
the error message will look like this:

	error: corrupt patch at .git/rebase-apply/patch:87

An alternate design was considered which involved printing the line
numbers relative to the output of `git am --show-current-patch` (in
other words, the actual mail file that's provided to am). This design
was not chosen because it is not always possible to give a line number
relative to the mbox file, say if the message were base64 encoded.

In addition, am does not store the whole mail and instead, splits the
mail into several files. As a result of this, this would break existing
users' workflow if they piped their mail directly to am from their mail
client, the whole mail would not exist in any file and they would have
to manually recreate the mail to see the line number.