t: use `set -o pipefail` on supported platforms
The current convention is to ensure that git commands are not placed in
the upstream of a pipe. If they are, they could fail in an undetectable
manner since a pipe's return code is the last command in the pipe.
However, many old tests are still written with git commands in the
upstream of a pipe.

In the spirit of catching these failures, run tests with
`set -o pipefail` if the underlying shell supports it. This way, we can
catch failures of Git commands that may occur even in the middle of a
pipeline.

Signed-off-by: Denton Liu <liu.denton@gmail.com>