t5004: ignore SIGPIPE in zipinfo
In a future patch, we plan on running tests with `set -o pipefail`.
Since zipinfo may be killed by SIGPIPE, it will return an error code
which will propogate as a result of the pipefail.

Mask away the return code of zipinfo so that the failure as a result of
the SIGPIPE does not propogate.

Signed-off-by: Denton Liu <liu.denton@gmail.com>