t5703: mask failing grep status
In a future patch, we plan on running tests with `set -o pipefail`.
However, since grep can return failure in the case that no lines are
matched, this can trigger a failure in a pipe in the case where grep is
being used as a filter.

Use test_fitler() instead of grep so that we don't fail even if we have
no matching lines.

Signed-off-by: Denton Liu <liu.denton@gmail.com>