t: define test_filter()
In a future patch, we plan on running tests with `set -o pipefail`.
However, since grep can return failure in the case that no lines are
matched, this can trigger a failure in a pipe in the case where grep is
being used as a filter.

Define the test_filter() function which acts as a wrapper around grep
but always returns 0 so that it can be used in the situation described
above.

Signed-off-by: Denton Liu <liu.denton@gmail.com>