t4205: cover `git log --reflog -z` blindspot
The test suite does not include any tests where `--reflog` and `-z` are
used together in `git log`. Cover this blindspot. Note that the
`--pretty=oneline` case is written separately because it follows a
slightly different codepath. Also, it is written as a one-iteration
for-loop because in a future patch, we will be testing more pretty
formats here.

Signed-off-by: Denton Liu <liu.denton@gmail.com>