t3701: make force_color() accept an optional prefix
In a future patch, we plan on making the test_must_fail()-family of
functions accept only git commands. Even though force_color() wraps an
invocation of `env git`, test_must_fail() will not be able to figure
this out since it will assume that force_color() is just some random
function which is disallowed.

Teach force_color() that if the first parameter is not "git", that it is
a prefix and run it before `env`. This is done because test_must_fail()
is a shell function and env cannot execute shell functions. However,
test_must_fail() may wrap env which allows us to write it like this.
This also futureproofs slightly because if a developer wants to run with
another prefix function, such as test_might_fail(), it should just work.