checkout: teach --worktree
A complaint that has come up frequently in the past is that it is not
possible to checkout files directly into a worktree without modifying
the index. Even though this could be worked around by redirecting the
output of `git show` to overwrite files, this was not feasible if one
wanted to use patch mode.

Since `git restore` was implemented, this has since been possible using
the `--worktree` option. However, some long-time users of Git still
prefer to use `git checkout` over `git restore` and would like to see
the functionality ported over.

Teach `git checkout --worktree`, allowing users to checkout files
directly into the worktree without affecting the index.