subtree: move test to root of project
The test for git subtree currently exists within its own directory. This
makes it difficult to catch any regressions that may occur as it is not
ran very often.

Move the test suite to the main test directory and guard it with the
GIT_SUBTREE prereq. This was a compromise taken that allows interested
parties to test git subtree more easily while not adding any additional
cost to those who don't care about it.