Documentation/Makefile: output to destination file dir
Documentation/Makefile correctly only generates documentation for files
under Documentation/. As a result of this, Makefiles under contrib/ have
to duplicate the logic contained here in order to generate their own
documentation. Let's avoid this and make the Makefile work when executed
on files under other directories.

$(XMLTO) only outputs files to Documentation/. Set the destination
directory so that it will correctly output to a directory when run on a
file outside of Documentation/.