Documentation/Makefile: only install if files present
Documentation/Makefile correctly only generates documentation for files
under Documentation/. As a result of this, Makefiles under contrib/ have
to duplicate the logic contained here in order to generate their own
documentation. Let's avoid this and make the Makefile work when executed
on files under other directories.

In a future commit, we will only be populating $(DOC_MAN1), leaving
$(DOC_MAN5) and $(DOC_MAN7) empty. If the Makefile attempts to install
with an empty file list, it will error out. Only attempt an $(INSTALL)
if the file list is non-empty.