contrib/subtree: make use of main Makefiles
The Makefile duplicates a lot of logic from the root Makefile and
Documentation/Makefile. Reduce this duplication of logic by calling
these Makefiles where possible. This allows this Makefile to stay more
up-to-date as it will automatically benefit from any updates to the
logic of the other Makefiles.

contrib/credential/netrc/Makefile was used as a template to make this
change.