Makefile: mark 'check-builtins' as a .PHONY target
The convention in Git's Makefile is to mark targets which do not
actually generate any files (such as static checking targets). .PHONY
enables the target to still run, even if a file is erroneously created
with the same name as the target.

Mark 'check-builtins' as a .PHONY target.

For interested readers, another convention that .PHONY targets should
follow is they should make use of the double-colon rule. This target
already follows that convention so we do not have to change it.