test-tool.h: ASCII-sort list of functions
The list of test-tool functions have, over time, gotten slightly out of
ASCII order. Sort this list to bring them back into order.

This happens to also be in alphabetical order. This happens to work out
because the ASCII value of '(' is less than '_' and the alphabetical
characters. This is unlike the sort of the Makefile where '.' comes
after '-' which results in a non-alphabetical sort.

This patch is best viewed with `--color-moved`.